Refactor to outer class
