+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/src/com/google/cloud/tools/eclipse/appengine/deploy/ui/internal/ProjectSelectorSelectionChangedListener.java
@@ -20,7 +20,6 @@ import com.google.api.client.auth.oauth2.Credential;
 import com.google.cloud.tools.eclipse.appengine.deploy.ui.Messages;
 import com.google.cloud.tools.eclipse.login.ui.AccountSelector;
 import com.google.cloud.tools.eclipse.projectselector.ProjectRepository;
-import com.google.cloud.tools.eclipse.projectselector.ProjectRepositoryException;
 import com.google.cloud.tools.eclipse.projectselector.ProjectSelector;
 import com.google.cloud.tools.eclipse.projectselector.model.AppEngine;
 import com.google.cloud.tools.eclipse.projectselector.model.GcpProject;
@@ -28,9 +27,6 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Predicate;
 import com.google.common.net.UrlEscapers;
 import java.text.MessageFormat;
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.jface.viewers.ISelectionChangedListener;
 import org.eclipse.jface.viewers.IStructuredSelection;
@@ -92,79 +88,5 @@ public class ProjectSelectorSelectionChangedListener implements ISelectionChange
           projectSelector, createAppLink, isLatestQueryJob, display);
       latestQueryJob.schedule();
     }
-  }
-}
-
-class AppEngineApplicationQueryJob extends Job {
-
-  private final GcpProject project;
-  private final Credential credential;
-  private final ProjectRepository projectRepository;
-  private final ProjectSelector projectSelector;
-  private final Predicate<Job> isLatestAppQueryJob;
-  private final String createAppLink;
-  private final Display display;
-
-  /**
-   * @param projectRepository {@link ProjectRepository#getAppEngineApplication} must be thread-safe
-   * @param isLatestQueryJob predicate that lazily determines if this job is the latest query job,
-   *     which determines if the job should update {@link ProjectSelector} or die silently. This
-   *     predicate is executed in the UI context
-   */
-  public AppEngineApplicationQueryJob(GcpProject project, Credential credential,
-      ProjectRepository projectRepository, ProjectSelector projectSelector, String createAppLink,
-      Predicate<Job> isLatestQueryJob, Display display) {
-    super("Checking GCP project has App Engine Application...");
-    this.project = project;
-    this.credential = credential;
-    this.projectRepository = projectRepository;
-    this.projectSelector = projectSelector;
-    this.createAppLink = createAppLink;
-    this.isLatestAppQueryJob = isLatestQueryJob;
-    this.display = display;
-  }
-
-  @Override
-  protected IStatus run(IProgressMonitor monitor) {
-    AppEngine appEngine = null;
-    String statusMessage = "";
-    String statusTooltip = null;
-
-    try {
-      // The original design was to cache the returned "appEngine" right after querying, but
-      // because we are in a non-UI thread, defer caching it until we get into the UI thread.
-      appEngine = projectRepository.getAppEngineApplication(credential, project.getId());
-
-      if (appEngine == AppEngine.NO_APPENGINE_APPLICATION) {
-        statusMessage = Messages.getString(
-            "projectselector.missing.appengine.application.link", createAppLink);
-        statusTooltip = createAppLink;
-      }
-    } catch (ProjectRepositoryException ex) {
-      statusMessage = Messages.getString(
-          "projectselector.retrieveapplication.error.message", ex.getLocalizedMessage());
-    }
-
-    if (appEngine != null || !statusMessage.isEmpty()) {
-      updateInUiThread(appEngine, statusMessage, statusTooltip);
-    }
-    return Status.OK_STATUS;
-  }
-
-  private void updateInUiThread(final AppEngine appEngine,
-      final String statusMessage, final String statusTooltip) {
-    final Job thisJob = this;
-
-    // The selector may have been disposed (i.e., dialog closed); check it in the UI thread.
-    display.syncExec(new Runnable() {
-      @Override
-      public void run() {
-        if (!projectSelector.isDisposed()
-            && isLatestAppQueryJob.apply(thisJob) /* intentionally checking in UI context */) {
-          project.setAppEngine(appEngine);
-          projectSelector.setStatusLink(statusMessage, statusTooltip);
-        }
-      }
-    });
   }
 }
